requirements.txt
# Core
anthropic>=0.40.0
synapseclient>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# HTTP
httpx>=0.27.0
requests>=2.31.0

# NCBI / Bioinformatics
biopython>=1.83

# Fuzzy matching / NLP
scikit-learn>=1.5.0

# Logging
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Dev utilities
python-dateutil>=2.9.0
